# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Naytia 像素小站 API",
  "description": "账号 / 论坛 / 工具箱收藏。契约见 docs/plans/03。",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/mcstructure/parse": {
   "post": {
    "tags": [
     "mcstructure"
    ],
    "summary": "上传 .mcstructure 并解析",
    "operationId": "parse_structure_api_mcstructure_parse_post",
    "parameters": [
     {
      "name": "include_voxels",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "是否返回完整方块索引数组",
       "default": false,
       "title": "Include Voxels"
      },
      "description": "是否返回完整方块索引数组"
     },
     {
      "name": "include_raw",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "是否返回完整根 NBT",
       "default": false,
       "title": "Include Raw"
      },
      "description": "是否返回完整根 NBT"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_parse_structure_api_mcstructure_parse_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StructureResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/mcstructure/voxels": {
   "post": {
    "tags": [
     "mcstructure"
    ],
    "summary": "只取完整方块索引数组",
    "operationId": "get_voxels_api_mcstructure_voxels_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_get_voxels_api_mcstructure_voxels_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/VoxelLayer"
         },
         "type": "array",
         "title": "Response Get Voxels Api Mcstructure Voxels Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/mcstructure/slice": {
   "post": {
    "tags": [
     "mcstructure"
    ],
    "summary": "取某一层薄片",
    "operationId": "get_slice_api_mcstructure_slice_post",
    "parameters": [
     {
      "name": "axis",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^[xyz]$",
       "description": "切片轴",
       "default": "y",
       "title": "Axis"
      },
      "description": "切片轴"
     },
     {
      "name": "at",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "description": "该轴上的坐标（结构内坐标，从 0 开始）",
       "title": "At"
      },
      "description": "该轴上的坐标（结构内坐标，从 0 开始）"
     },
     {
      "name": "layer",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "层号：0 主层 / 1 次层",
       "default": 0,
       "title": "Layer"
      },
      "description": "层号：0 主层 / 1 次层"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_get_slice_api_mcstructure_slice_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VoxelSlice"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/mcstructure/position": {
   "post": {
    "tags": [
     "mcstructure"
    ],
    "summary": "坐标 ↔ 位置下标互查",
    "operationId": "lookup_position_api_mcstructure_position_post",
    "parameters": [
     {
      "name": "x",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "description": "结构内 X（从 0 开始）",
       "title": "X"
      },
      "description": "结构内 X（从 0 开始）"
     },
     {
      "name": "y",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "description": "结构内 Y（从 0 开始）",
       "title": "Y"
      },
      "description": "结构内 Y（从 0 开始）"
     },
     {
      "name": "z",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "description": "结构内 Z（从 0 开始）",
       "title": "Z"
      },
      "description": "结构内 Z（从 0 开始）"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_lookup_position_api_mcstructure_position_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PositionLookup"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/mcstructure/blocks": {
   "post": {
    "tags": [
     "mcstructure"
    ],
    "summary": "方块实体 / 实体分页列表",
    "operationId": "list_blocks_api_mcstructure_blocks_post",
    "parameters": [
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(block_entities|entities)$",
       "description": "block_entities = 方块实体；entities = 实体",
       "default": "block_entities",
       "title": "Kind"
      },
      "description": "block_entities = 方块实体；entities = 实体"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_list_blocks_api_mcstructure_blocks_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BlockEntityPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BlockCountEntry": {
    "properties": {
     "index": {
      "type": "integer",
      "title": "Index",
      "description": "调色板下标"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "states": {
      "type": "object",
      "title": "States"
     },
     "count": {
      "type": "integer",
      "title": "Count",
      "description": "出现次数（含所有层）"
     },
     "ratio": {
      "type": "number",
      "title": "Ratio",
      "description": "占非 void 方块总数的比例，0~1"
     }
    },
    "type": "object",
    "required": [
     "index",
     "name",
     "count",
     "ratio"
    ],
    "title": "BlockCountEntry"
   },
   "BlockEntityEntry": {
    "properties": {
     "index": {
      "type": "integer",
      "title": "Index",
      "description": "位置下标（ZYX 顺序）；用 /position 可换算成坐标"
     },
     "x": {
      "type": "integer",
      "title": "X"
     },
     "y": {
      "type": "integer",
      "title": "Y"
     },
     "z": {
      "type": "integer",
      "title": "Z"
     },
     "identifier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Identifier",
      "description": "block_entity_data.id"
     },
     "block_entity_data": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Block Entity Data",
      "description": "方块实体 NBT 原文（箱子内容、告示牌文字等）"
     },
     "tick_queue_data": {
      "items": {},
      "type": "array",
      "title": "Tick Queue Data",
      "description": "计划刻队列（如珊瑚死亡、水流更新）"
     }
    },
    "type": "object",
    "required": [
     "index",
     "x",
     "y",
     "z"
    ],
    "title": "BlockEntityEntry",
    "description": "block_position_data 的一条：方块实体 NBT 与计划刻队列。"
   },
   "BlockEntityPage": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "items": {
      "items": {
       "type": "object"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "total",
     "offset",
     "limit",
     "items"
    ],
    "title": "BlockEntityPage",
    "description": "方块实体 / 实体列表的分页结果（大结构里可能有上万条）。"
   },
...
   "BlockStateEntry": {
    "properties": {
     "index": {
      "type": "integer",
      "title": "Index",
      "description": "在调色板中的下标；block_indices 里的值就是它"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "description": "方块标识，如 minecraft:stone"
     },
     "states": {
      "type": "object",
      "title": "States",
      "description": "方块状态。值类型随状态而定：字符串 / 整数 / 字节（布尔）"
     },
     "version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Version",
      "description": "该方块的兼容版本号"
     }
    },
    "type": "object",
    "required": [
     "index",
     "name"
    ],
    "title": "BlockStateEntry",
    "description": "调色板里的一个方块排列（block_palette 的一项）。"
   },
...
   "Body_get_slice_api_mcstructure_slice_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File",
      "description": ".mcstructure 文件"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_get_slice_api_mcstructure_slice_post"
   },
   "Body_get_voxels_api_mcstructure_voxels_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File",
      "description": ".mcstructure 文件"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_get_voxels_api_mcstructure_voxels_post"
   },
   "Body_list_blocks_api_mcstructure_blocks_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File",
      "description": ".mcstructure 文件"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_list_blocks_api_mcstructure_blocks_post"
   },
   "Body_lookup_position_api_mcstructure_position_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File",
      "description": ".mcstructure 文件"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_lookup_position_api_mcstructure_position_post"
   },
   "Body_parse_structure_api_mcstructure_parse_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File",
      "description": ".mcstructure 文件"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_parse_structure_api_mcstructure_parse_post"
   },
...
   "EntityEntry": {
    "properties": {
     "index": {
      "type": "integer",
      "title": "Index",
      "description": "在 entities 列表中的序号"
     },
     "block_position": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StructureSize"
       },
       {
        "type": "null"
       }
      ],
      "description": "保存时的方块坐标（相对结构）；Pos 是绝对世界坐标，加载时会被替换"
     },
     "identifier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Identifier",
      "description": "实体标识，如 minecraft:armor_stand"
     },
     "data": {
      "type": "object",
      "title": "Data",
      "description": "实体 NBT 原文，含 Pos / UniqueID / 自定义字段"
     }
    },
    "type": "object",
    "required": [
     "index"
    ],
    "title": "EntityEntry",
    "description": "structure.entities 的一条：完整实体 NBT。"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LayerStats": {
    "properties": {
     "layer": {
      "type": "integer",
      "title": "Layer",
      "description": "0 = 主层，1 = 次层（共位方块，如水下的水+水草）"
     },
     "filled": {
      "type": "integer",
      "title": "Filled",
      "description": "该层中非 void 的方块数"
     },
     "void": {
      "type": "integer",
      "title": "Void",
      "description": "该层中 void(-1) 的数量"
     },
     "fill_ratio": {
      "type": "number",
      "title": "Fill Ratio",
      "description": "filled / 总格数，0~1"
     }
    },
    "type": "object",
    "required": [
     "layer",
     "filled",
     "void",
     "fill_ratio"
    ],
    "title": "LayerStats"
   },
...
   "PositionLookup": {
    "properties": {
     "x": {
      "type": "integer",
      "title": "X"
     },
     "y": {
      "type": "integer",
      "title": "Y"
     },
     "z": {
      "type": "integer",
      "title": "Z"
     },
     "index": {
      "type": "integer",
      "title": "Index",
      "description": "位置下标 = x*(sy*sz) + y*sz + z"
     },
     "size": {
      "$ref": "#/components/schemas/StructureSize"
     },
     "layers": {
      "items": {
       "type": "object"
      },
      "type": "array",
      "title": "Layers",
      "description": "每层在这一格的方块：layer 与调色板下标；-1 表示 void"
     }
    },
    "type": "object",
    "required": [
     "x",
     "y",
     "z",
     "index",
     "size"
    ],
    "title": "PositionLookup",
    "description": "坐标 <-> 位置下标互查结果。"
   },
...
   "StructureLayout": {
    "properties": {
     "size": {
      "$ref": "#/components/schemas/StructureSize"
     },
     "voxel_count": {
      "type": "integer",
      "title": "Voxel Count",
      "description": "size 的乘积"
     },
     "world_origin": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StructureSize"
       },
       {
        "type": "null"
       }
      ],
      "description": "structure_world_origin：保存时的世界原点"
     },
     "world_origin_source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "World Origin Source",
      "description": "原点取自哪里。真实文件放在根层级（'root'），bedrock.dev 文档写在 structure 内部（'structure'）；缺失时为 null"
     },
     "coordinate_order": {
      "type": "string",
      "title": "Coordinate Order",
      "description": "位置下标的排列顺序；换算公式见 /position 接口说明",
      "default": "zyx"
     }
    },
    "type": "object",
    "required": [
     "size",
     "voxel_count"
    ],
    "title": "StructureLayout"
   },
...
   "StructureResponse": {
    "properties": {
     "layout": {
      "$ref": "#/components/schemas/StructureLayout"
     },
     "format_version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Format Version"
     },
     "compression": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compression",
      "description": "检测到的压缩方式：null / gzip / zlib"
     },
     "file_bytes": {
      "type": "integer",
      "title": "File Bytes",
      "description": "上传文件字节数"
     },
     "palette": {
      "items": {
       "$ref": "#/components/schemas/BlockStateEntry"
      },
      "type": "array",
      "title": "Palette"
     },
     "block_entities": {
      "items": {
       "$ref": "#/components/schemas/BlockEntityEntry"
      },
      "type": "array",
      "title": "Block Entities"
     },
     "entities": {
      "items": {
       "$ref": "#/components/schemas/EntityEntry"
      },
      "type": "array",
      "title": "Entities"
     },
     "stats": {
      "$ref": "#/components/schemas/StructureStats"
     },
     "extra_root_fields": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Extra Root Fields",
      "description": "未建模但已保留的根字段名（不丢数据）"
     },
     "raw_nbt": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Raw Nbt",
      "description": "完整根 NBT（?include_raw=true 时返回）"
     },
     "layer_sizes": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Layer Sizes",
      "description": "各层的格数，供 /voxels 分页判断"
     },
     "voxels": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/VoxelLayer"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voxels",
      "description": "完整方块索引。默认不返回；?include_voxels=true 时返回"
     }
    },
    "type": "object",
    "required": [
     "layout",
     "file_bytes",
     "palette",
     "block_entities",
     "entities",
     "stats",
     "extra_root_fields",
     "layer_sizes"
    ],
    "title": "StructureResponse",
    "description": "/parse 的响应：索引 + 可选的完整体素数组。"
   },
   "StructureSize": {
    "properties": {
     "x": {
      "type": "integer",
      "title": "X",
      "description": "X 方向尺寸（方块数）"
     },
     "y": {
      "type": "integer",
      "title": "Y",
      "description": "Y 方向尺寸（方块数）"
     },
     "z": {
      "type": "integer",
      "title": "Z",
      "description": "Z 方向尺寸（方块数）"
     }
    },
    "type": "object",
    "required": [
     "x",
     "y",
     "z"
    ],
    "title": "StructureSize"
   },
   "StructureStats": {
    "properties": {
     "total_voxels": {
      "type": "integer",
      "title": "Total Voxels",
      "description": "size 的乘积，等于每层的格数"
     },
     "filled": {
      "type": "integer",
      "title": "Filled",
      "description": "非 void 方块总数（含两层，可能大于格数）"
     },
     "void_slots": {
      "type": "integer",
      "title": "Void Slots",
      "description": "主层的 void 数量"
     },
     "primary_filled": {
      "type": "integer",
      "title": "Primary Filled"
     },
     "secondary_filled": {
      "type": "integer",
      "title": "Secondary Filled"
     },
     "out_of_range_indices": {
      "type": "integer",
      "title": "Out Of Range Indices",
      "description": "越界调色板下标数量。游戏加载时按空气处理，这里如实计数",
      "default": 0
     },
     "palette_used": {
      "type": "integer",
      "title": "Palette Used",
      "description": "实际被引用到的调色板条目数"
     },
     "blocks": {
      "items": {
       "$ref": "#/components/schemas/BlockCountEntry"
      },
      "type": "array",
      "title": "Blocks",
      "description": "按数量降序"
     },
     "layers": {
      "items": {
       "$ref": "#/components/schemas/LayerStats"
      },
      "type": "array",
      "title": "Layers"
     }
    },
    "type": "object",
    "required": [
     "total_voxels",
     "filled",
     "void_slots",
     "primary_filled",
     "secondary_filled",
     "palette_used",
     "blocks",
     "layers"
    ],
    "title": "StructureStats"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VoxelLayer": {
    "properties": {
     "layer": {
      "type": "integer",
      "title": "Layer"
     },
     "indices": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Indices",
      "description": "长度 = size 的乘积，按 ZYX 顺序（从底部西北角到顶部东南角）"
     }
    },
    "type": "object",
    "required": [
     "layer",
     "indices"
    ],
    "title": "VoxelLayer",
    "description": "一层的方块索引数组（值 = 调色板下标，-1 表示 void）。"
   },
   "VoxelSlice": {
    "properties": {
     "axis": {
      "type": "string",
      "title": "Axis",
      "description": "切片轴：x / y / z"
     },
     "layer": {
      "type": "integer",
      "title": "Layer",
      "description": "层号：0 主层 / 1 次层"
     },
     "at": {
      "type": "integer",
      "title": "At",
      "description": "该轴上的坐标"
     },
     "plane_size": {
      "$ref": "#/components/schemas/StructureSize",
      "description": "薄片的尺寸（该轴恒为 1）"
     },
     "indices": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Indices",
      "description": "按 ZYX 顺序展开的调色板下标；-1 表示 void"
     },
     "blocks": {
      "items": {
       "$ref": "#/components/schemas/BlockCountEntry"
      },
      "type": "array",
      "title": "Blocks",
      "description": "该薄片的方块统计"
     }
    },
    "type": "object",
    "required": [
     "axis",
     "layer",
     "at",
     "plane_size",
     "indices"
    ],
    "title": "VoxelSlice",
    "description": "一次薄片查询的结果。坐标为结构内坐标（从 0 开始）。"
   },
...
  }
 }
}